make Property a function creating a PropertySpec instance
PropertySpec is what was called Property. So this commit simply delegates its instantiation to a function, which is now called Property. By doing so, one can suppress type checkers' error noting assignment of "Property" instance to the variables annotated with real types. For most users that use what was called Property to create its instances, this does not cause any runtime effect.

Despite this, it IS a breaking change for a (presumably) small portion of users assuming Property is a type. For example, if one subclasses Property or inspects types against Property, they need to replace it with PropertySpec.